requirements.txt
requests
numpy
scipy
scikit-learn
scikit-plot
certifi
chardet
Click
cloudpickle
cycler
idna
kiwisolver
matplotlib
pandas
Pillow
pyparsing
python-dateutil
pytz
six
torch
torchvision
tqdm
ujson
urllib3
